progress

+++ projects/android/classygames/src/edu/selu/android/classygames/games/checkers/Board.java
@@ -36,7 +36,7 @@ public class Board extends GenericBoard
 	@Override
 	public boolean isPositionValid(final Coordinate coordinate)
 	{
-		return isPositionValid((byte) coordinate.getX(), (byte) coordinate.getY());
+		return isPositionValid(coordinate.getX(), coordinate.getY());
 	}
